Update non-decimals fields

+++ app/v2/models/response/getAllowancesDeductionsAndReliefs/detail/ForeignPropertyRfcDetail.scala
@@ -19,7 +19,7 @@ package v2.models.response.getAllowancesDeductionsAndReliefs.detail
 import play.api.libs.json.{Json, OFormat}
 
 case class ForeignPropertyRfcDetail(countryCode: String,
-                                    amountClaimed: BigDecimal,
+                                    amountClaimed: BigInt,
                                     allowableAmount: BigDecimal,
                                     carryForwardAmount: Option[BigDecimal])
